adding (unsigned long int number)to the main.h

+++ main.h
@@ -18,7 +18,7 @@ int print_octal(unsigned int number);
 int print_hex(unsigned int number);
 int print_HEX(unsigned int number);
 
-int print_pointer(unsigned int number);
+int print_pointer(unsigned long int number);
 
 int print_reversed(va_list arg);
 int print_R(va_list arg);
